+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [0.0.6]
+### Added
+  - Error message popup for any errors that were encountered
+### Changed
+  - If there are extra custom templates that has the known file suffixes (component, service, pipe, etc..), they will be rendered as well.
+
 ## [0.0.5]
 ### Added
   - More generator options: Route, Directive, Pipe